Move backends setup files
