+++ mne/viz/utils.py
@@ -15,8 +15,6 @@ from functools import partial
 import difflib
 import webbrowser
 import tempfile
-import os
-import importlib
 import numpy as np
 from copy import deepcopy
 from distutils.version import LooseVersion
@@ -44,46 +42,6 @@ _channel_type_prettyprint = {'eeg': "EEG channel", 'grad': "Gradiometer",
                              'eog': "EOG channel", 'ecg': "ECG sensor",
                              'emg': "EMG sensor", 'ecog': "ECoG channel",
                              'misc': "miscellaneous sensor"}
-
-default_3D_backend = 'mlab'
-
-
-global MNE_3D_BACKEND
-try:
-    if MNE_3D_BACKEND is None:
-        MNE_3D_BACKEND = os.environ.get('MNE_3DMNE_3D_BACKEND',
-                                        default_3D_backend)
-except NameError:
-    MNE_3D_BACKEND = os.environ.get('MNE_3DMNE_3D_BACKEND', default_3D_backend)
-
-
-def set_3D_backend(backend_name):
-    """Set the backend for MNE.
-
-    The backend will be set as specified and operations will use
-    that backend
-
-    Parameters
-    ----------
-    backend_name : {'mlab'}, default is 'mlab'
-    """
-    global MNE_3D_BACKEND
-    MNE_3D_BACKEND = backend_name
-    from . import backend
-    importlib.reload(backend)
-
-
-def get_3D_backend():
-    """Return the backend currently used.
-
-    Returns
-    -------
-    backend_used : str
-        the backend currently in use
-    """
-    global MNE_3D_BACKEND
-    backend_used = MNE_3D_BACKEND
-    return backend_used
 
 
 def _setup_vmin_vmax(data, vmin, vmax, norm=False):
